Move rename module `agent` -> `agents`
