+++ autogpt/agents/__init__.py
@@ -1,0 +1,4 @@
+from .agent import Agent
+from .agent_manager import AgentManager
+
+__all__ = ["Agent", "AgentManager"]
