+++ autogpt/commands/web_search.py
@@ -7,7 +7,7 @@ from itertools import islice
 
 from duckduckgo_search import DDGS
 
-from autogpt.agent.agent import Agent
+from autogpt.agents.agent import Agent
 from autogpt.command_decorator import command
 
 DUCKDUCKGO_MAX_ATTEMPTS = 3
