+++ tests/unit/test_git_commands.py
@@ -2,7 +2,7 @@ import pytest
 from git.exc import GitCommandError
 from git.repo.base import Repo
 
-from autogpt.agent.agent import Agent
+from autogpt.agents.agent import Agent
 from autogpt.commands.git_operations import clone_repository
